requirements.txt
streamlit
flask
pillow
scikit-learn
ollama
requests
pytesseract
opencv-python
colorthief
